Create README.md
# Atari's Pong Game

This is a replica of the original Atari's Pong game made in Unity by me. The idea was to keep all the features just like the original game. To a better understanding of the code, you may study a bit of normal vectors and reflection vectors. There is no need to be an expert in Linear Algebra to make this game, however some of its concepts are crucial for game developers.